backend/requirements.txt
paddlepaddle
PaddleOCR
fastapi
numpy
Pillow
pydantic
ultralytics
uvicorn
python-multipart